sampling training data

+++ src/ModelOps/model_fit.py
@@ -45,7 +45,7 @@ class ModelFit:
             )
         #create a sample from the training data
         PERCENTAGE=0.2
-        ref_dataset=train.sample(PERCENTAGE*train.shape[0],replace=True)
+        ref_dataset=train.sample(int(PERCENTAGE*train.shape[0]),replace=True)
         os.makedirs("artifacts/reference_data",exist_ok=True)
         ref_dataset.to_csv('artifacts/reference_data/output.csv',index=False)
